requirements.txt
python-telegram-bot==20.3
openai>=1.0.0
psycopg2-binary
python-dotenv
pydantic>=2.0.0